Circular import hotfix
Attempt to fix circular imports by moving reconnect_on_fail to a new
file utility.py. Can't test code at all so failure is quite possible.
